Move deployment tests to InappropriateDeploymentTest
